+++ src/export.rs
@@ -13,12 +13,20 @@ pub use embassy_executor::task;
 
 const VERSION: u32 = 1; //Format version of our protocol between probe-rs and target running embedded-test
 
+// Constants ahead that are used by proc macros to calculate the needed json buffer size
+pub const JSON_SIZE_HEADER : usize = r#"{"version:12,tests:[]}"#.len();
+pub const JSON_SIZE_PER_TEST_WITHOUT_TESTNAME : usize = r#"{"name":"","should_panic":false,ignored:false,timeout:1234567890},"#.len();
+
+/// Struct which will be serialized as JSON and sent to probe-rs when it requests the available tests
+// NOTE: Update the const's above if you change something here
 #[derive(Debug, serde::Serialize)]
-pub struct Tests<'a> {
+struct Tests<'a> {
     pub version: u32,
     pub tests: &'a [Test],
 }
 
+/// Struct which describes a single test to probe-rs
+// NOTE: Update the const's above if you change something here
 #[derive(Debug, serde::Serialize)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct Test {
@@ -30,14 +38,14 @@ pub struct Test {
     pub timeout: Option<u32>,
 }
 
-pub fn run_tests(tests: &mut [Test]) -> ! {
+pub fn run_tests<const JSON_SIZE_TOTAL: usize>(tests: &mut [Test]) -> ! {
     for test in tests.iter_mut() {
         test.name = &test.name[test.name.find("::").unwrap() + 2..];
     }
 
-    let mut args = &semihosting::experimental::env::args::<1024>().unwrap(); //TODO: handle error
+    let mut args = &semihosting::experimental::env::args::<1024>().expect("Failed to get cmdline via semihosting");
     let command = match args.next() {
-        Some(c) => c.unwrap(),
+        Some(c) => c.expect("command to run contains non-utf8 characters"),
         None => {
             error!("Received no arguments via semihosting. Please communicate with the target with the embedded-test runner.");
             semihosting::process::abort();
@@ -52,8 +60,8 @@ pub fn run_tests(tests: &mut [Test]) -> ! {
                 tests,
             };
 
-            let mut buf = [0u8; 1024];
-            let size = serde_json_core::to_slice(&tests, &mut buf).unwrap();
+            let mut buf = [0u8; JSON_SIZE_TOTAL];
+            let size = serde_json_core::to_slice(&tests, &mut buf).expect("Buffer to store list of test was too small");
             let args = [ParamRegR::ptr(buf.as_ptr()), ParamRegR::usize(size)];
             let ret = unsafe {
                 syscall_readonly(
@@ -69,8 +77,8 @@ pub fn run_tests(tests: &mut [Test]) -> ! {
             semihosting::process::exit(0);
         }
         "run" => {
-            let test_name = args.next().unwrap().unwrap();
-            let test = tests.iter_mut().find(|t| t.name == test_name).unwrap();
+            let test_name = args.next().expect("test name missing").expect("test name contains non-utf8 character");
+            let test = tests.iter_mut().find(|t| t.name == test_name).expect("test to run not found");
             info!("Running test: {:?}", test);
             (test.function)();
         }